Increase document size to 16mb #1295
