+++ LiteDB.Tests/Expressions/Expressions_Exec_Tests.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using FluentAssertions;
+using System;
 using System.Collections.Generic;
 using Xunit;
 
@@ -328,7 +329,7 @@ namespace LiteDB.Tests.Expressions
         }
 
         [Fact]
-        public void Multiple_And_Tests()
+        public void Expression_Multiple_And_Tests()
         {
             BsonValue B(params BsonExpression[] args)
             {
@@ -342,7 +343,7 @@ namespace LiteDB.Tests.Expressions
         }
 
         [Fact]
-        public void Multiple_Or_Tests()
+        public void Expression_Multiple_Or_Tests()
         {
             BsonValue B(params BsonExpression[] args)
             {
@@ -353,6 +354,40 @@ namespace LiteDB.Tests.Expressions
             B("5 < 3", "10 != 'a'", "1 = 1").ExpectValue(true);
             B("5 > 3", "10 = 'a'", "1 = 1").ExpectValue(true);
             B("5 < 3", "10 = 'a'", "1 != 1").ExpectValue(false);
+        }
+
+        [Fact]
+        public void Expression_AndAlso_OrElse()
+        {
+            var ex1 = BsonExpression.Create("LENGTH($.x) >= 5 AND SUBSTRING($.x, 0, 5) = \"12345\"");
+            var doc1 = new BsonDocument();
+
+            // OK (true)
+            doc1["x"] = "12345";
+            ex1.ExecuteScalar(doc1);
+
+            // KO (expected: false, actual: exception)
+            doc1["x"] = "123";
+            ex1.ExecuteScalar(doc1);
+        }
+
+        [Fact]
+        public void Expression_Conditional_IIF()
+        {
+            var ex1 = BsonExpression.Create("IIF(LENGTH($.x) >= 5, SUBSTRING($.x, 0, 5), \"too-short\")");
+            var doc1 = new BsonDocument();
+
+            // OK ("12345")
+            doc1["x"] = "123456789";
+            var r1 = ex1.ExecuteScalar(doc1);
+
+            r1.AsString.Should().Be("12345");
+
+            // KO (expected: "too-short", actual: System.ArgumentOutOfRangeException: Index and length must refer to a location within the string.)
+            doc1["x"] = "123";
+            var r2 = ex1.ExecuteScalar(doc1);
+
+            r2.AsString.Should().Be("too-short");
         }
     }
 }
